Changed amount awarded to float field

+++ datamad2/models/grants.py
@@ -116,7 +116,7 @@ class ImportedGrant(models.Model):
     grant_status = models.CharField(max_length=50, default="Active",
                                     choices=(("Active", "Active"), ("Closed", "Closed")))
     # AmountAwarded	Siebel	Amount in pounds stirling			AMOUNT
-    amount_awarded = models.DecimalField(null=True, blank=True, max_digits=10, decimal_places=2)
+    amount_awarded = models.FloatField(null=True, blank=True)
     # Call	Siebel	E.g. Standard Grant DEC06			CALL
     # ignore call for now
     call = models.CharField(max_length=1024, default='', blank=True)
